starting to refactor tests
