+++ tests/test_getconfig.py
@@ -1,20 +1,10 @@
 from chaos_lambda import get_config
 from ssm_cache import InvalidParameterError
+from test_abstract import client, ignore_warnings
 import unittest
 import os
 import warnings
 import boto3
-
-client = boto3.client('ssm', region_name='eu-north-1')
-
-
-def ignore_warnings(test_func):
-    def do_test(self, *args, **kwargs):
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", ResourceWarning)
-            warnings.simplefilter("ignore", DeprecationWarning)
-            test_func(self, *args, **kwargs)
-    return do_test
 
 
 class TestConfigMethods(unittest.TestCase):
