+++ tests/test_session.py
@@ -1,19 +1,11 @@
 from chaos_lambda import SessionWithDelay
 from unittest.mock import patch
+from test_abstract import client, ignore_warnings
 from io import StringIO
 import unittest
 import warnings
 import pytest
 import logging
-
-
-def ignore_warnings(test_func):
-    def do_test(self, *args, **kwargs):
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", ResourceWarning)
-            warnings.simplefilter("ignore", DeprecationWarning)
-            test_func(self, *args, **kwargs)
-    return do_test
 
 
 def session_request_with_delay():
